LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-dropping-controls.html


<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-service.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/start-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/play-pause-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/skip-back-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/airplay-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/pip-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/fullscreen-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/mute-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/tracks-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/slider.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/volume-slider.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scrubber.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/time-label.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/time-control.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/placard.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/airplay-placard.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/pip-placard.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/buttons-container.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/media-controls.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/macos-media-controls.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/macos-inline-media-controls.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing <code>MacOSInlineMediaControls</code> and dropping controls.");

window.jsTestIsAsync = true;

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const mediaControls = new MacOSInlineMediaControls;

// There is a 4-pixel difference in metrics between 1x and 2x due to the volume button having a different width.
const is1x = window.devicePixelRatio == 1;
const expectedWidthsForDrop = [
    [mediaControls.timeControl, is1x ? 655 : 651],
    [mediaControls.airplayButton, is1x ? 394 : 390],
    [mediaControls.pipButton, is1x ? 345 : 341],
    [mediaControls.tracksButton, is1x ? 294 : 290],
    [mediaControls.muteButton, is1x ? 245 : 241],
    [mediaControls.skipBackButton, 187],
    [mediaControls.fullscreenButton, 139]
];

function dumpControlVisibility(control)
{
    let visibilityString = "visible";
    if (!control.parent)
        visibilityString = "not " + visibilityString;
    debug(`${control.constructor.name} is ${visibilityString}`);
}

const iconButtons = [
    mediaControls.playPauseButton,
    mediaControls.skipBackButton,
    mediaControls.airplayButton,
    mediaControls.pipButton,
    mediaControls.tracksButton,
    mediaControls.muteButton,
    mediaControls.fullscreenButton
];

scheduler.frameDidFire = function()
{
    if (iconButtons.some(button => button.width == 0))
        return;

    for (let [control, width] of expectedWidthsForDrop) {
        mediaControls.width = width + 1;
        mediaControls.layout();
        dumpControlVisibility(control);

        mediaControls.width = width;
        mediaControls.layout();
        dumpControlVisibility(control);

        debug("");
    }

    scheduler.frameDidFire = new Function;
    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
